Call into libsasl to check Basic auth passwords
Use libsasl2 to check passwords, offering it as an option for Basic
authentication with a configurable realm name.  The SASL service name
used for the password check is hard-coded to "docker".

Deployments which expect to use saslauthd for checking the passwords
will need to create a file named "docker.conf" under /usr/lib/sasl2 or
/etc/sasl2 containing this line:
  pwcheck_method:saslauthd

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>