Call into gssapi to check Negotiate auth tokens
Use the GSSAPI libraries to accept GSSAPI tokens, and offer the option
of using it as the Negotiate mechanism if we can find server credentials
to use at startup-time.

If --keytab is passed as a command-line option, we'll attempt to
register that as the Kerberos keytab to be used, also at startup-time.

We don't set a service name, so we'll accept authentication as any
service for which keys can be found in the keytab, but the expectation
is that clients will authenticate to the service named "HTTP" on the
host, since that's what the Negotiate spec says they should use.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>