Ask the kernel to identify local daemon clients
On Linux systems, ask the kernel for the Uid on the other end of a Unix
socket connection, and look up the name of the user.  This provides an
implicit authentication for the common cases.

Note that the lookup is currently only supported for non-static builds,
because we use the standard library's lookup logic, which only returns
an error in static builds.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>